Working version of code that checks the sum rule
This code is based on first finding the peak of the omega-integrand, and
using that information to perform an adaptive integration. This seems to
work fairly well in the best cases, but is by no means a universal or
'hands-off' code. Tinkering still needs to be done in some cases.

Currently, I am only using omegaintegral_check to first check how well
the omega integral performed by eye + sum rules, and then I'll do the
k-integral using the outputs from this function.