Improvements to ELFmax finder function
In some cases, the ELFmax function fails. I tried to remedy at
least one case where the failure occurs, namely when the ELF
returns nan (i.e. we hit an actual delta function). I also changed
the code so it doesn't stop the rest of the process if the ELFmax
function fails; the code now reports a failure to reach the sum
rule for the current k point and move onto the next k point.